improve performance by using int for increment in for-cycles

+++ src/main/java/matrix/Modulo2ArraysMultiplier.java
@@ -3,10 +3,10 @@ package matrix;
 public class Modulo2ArraysMultiplier implements ArraysMultiplier {
     @Override
     public byte multiply(short arraysSize, byte[] array1, byte[] array2) {
-        byte result = 0;
-        for (short i=0; i<arraysSize; i++) {
+        int result = 0;
+        for (int i=0; i<arraysSize; i++) {
             result ^= array1[i] & array2[i];
         }
-        return result;
+        return (byte) result;
     }
 }
